+++ app/__init__.py
@@ -1,0 +1,20 @@
+from flask import Flask
+import config
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import LoginManager
+
+
+
+app = Flask(__name__)
+app.config.from_object(config)
+db = SQLAlchemy(app)
+lm = LoginManager()
+lm.login_view = 'login'
+
+
+
+from app import views
+import models
+
+if __name__ == '__main__':
+    app.run(debug=True)
