+++ app/templates/article_edit.html
@@ -1,0 +1,58 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.css">
+    <link href="//fonts.gdgdocs.org/css?family=Lobster" rel="stylesheet" type="text/css">
+    <style>
+        h1 { font-family: 'Pacifico', cursive; }
+        hr { width: 100px; }
+    </style>
+    <title>Goodwriter</title>
+</head>
+<body>
+    <div class="container-fluid">
+        <h1 class="text-center">Goodwriter</h1>
+
+    <hr/>
+        <div class="row">
+            <div class="col-sm-3">
+
+                <div class="well">
+                  <input type="text" placeholder="输入查找内容">
+                  <button type="submit" class="btn btn-info btn-xs">搜索</button>
+                    <form class="form-inline">
+                        <div >
+                            <button class="btn btn-link btn-lg"><span class="glyphicon glyphicon-paperclip" aria-hidden="true">标签</span></button>
+                              <br/>
+                            <button type="button" class="btn btn-link btn-lg">电影</button>
+
+                            <button type="button" class="btn btn-link btn-lg">音乐</button>
+
+                            <button type="button" class="btn btn-link btn-lg">python</button>
+
+                            <button type="button" class="btn btn-link btn-lg">感悟</button>
+
+                        </div>
+                    </form>
+
+                </div>
+                <div class="well" id="result">
+                  这里用来放查询结果
+                </div>
+            </div>
+            <div class="col-sm-9">
+              <h3>标签</h3>
+              <input type="text" class="form-control" placeholder="标题" name="tag">
+              <h3>内容</h3>
+              <input type="text" class="form-control" placeholder="内容" name="context">
+
+              <button class="btn btn-primary" type="submit">提交</button>
+
+            </div>
+        </div>
+</body>
+</html>
