+++ app/templates/login.html
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.css">
+    <link href="//fonts.gdgdocs.org/css?family=Lobster" rel="stylesheet" type="text/css">
+    <style>
+        h1 { font-family: 'Pacifico', cursive; }
+        h2 { font-family: 'Pacifico', cursive; }
+        h3 { font-family: 'Pacifico', cursive; }
+        p { font-family: 'Pacifico', cursive; }
+        a { font-family: 'Pacifico', cursive;color: black; }
+        label { font-family: 'Pacifico', cursive; }
+    </style>
+    <title>Goodwriter</title>
+</head>
+<body>
+    <div class="container-fluid">
+        <div class="page-header">
+            <h1 class="text-center">Goodwriter</h1>
+        </div>
+    </div>
+    <div class="container-fluid">
+        <div class="col-md-2">
+            <div>
+                <h2>Goodwriter</h2>
+            </div>
+            <div>
+                <h4><a href="#">登录</a> | <a href="/regist">注册</a></h4>
+            </div>
+        </div>
+        <div class="col-md-1">
+        </div>
+        <div class="col-md-4">
+          <form action="" method="POST">
+            <div>
+                <h2>
+                    登录 或 <a href="regist"><span>注册</span></a>
+                </h2>
+            </div>
+            <hr>
+            <div class="btn-group pull-center">
+                <h3>Email</h3>
+                <input type="text" class="form-control" placeholder="example@163.com" name="email">
+                <h3>Password</h3>
+                <input type="password" class="form-control" placeholder="Password" name="password">
+                <input type="checkbox">
+                <label>Remember me?</label>
+            </div>
+            <div>
+                <button class="btn btn-primary" type="submit">登录</button>
+            </div>
+            <div>
+                <a href="#">Forget password?</a>
+            </div>
+          </form>
+        </div>
+    </div>
+</body>
+</html>
