+++ app/templates/material_list.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://cdn.bootcss.com/bootstrap/3.3.7/css/bootstrap.css">
+    <link href="//fonts.gdgdocs.org/css?family=Lobster" rel="stylesheet" type="text/css">
+    <style>
+        h1 { font-family: 'Pacifico', cursive; }
+        hr { width: 100px; }
+    </style>
+    <title>Goodwriter</title>
+</head>
+<body>
+    <div class="container-fluid">
+        <h1 class="text-center">Goodwriter</h1>
+        <div class="row" class="col-sm-6">
+            <div class="col-md-6">
+                <input type="text" placeholder="输入查找内容">
+                <button type="submit" class="btn btn-info btn-xs">搜索</button>
+            </div>
+            <div class="col-md-6">
+                <button type="submit" class="btn btn-default btn-sm pull-right RbtnMargin">写文章</button>
+                <button type="submit" class="btn btn-default btn-sm pull-right">写素材</button>
+            </div>
+        </div>
+    <hr>
+        <div class="row">
+            <div class="col-sm-3">
+                <div class="well">
+                    <form class="form-inline">
+                        <div class="btn-group-vertical" role="group">
+                            <button class="btn btn-link btn-lg"><span class="glyphicon glyphicon-paperclip" aria-hidden="true">标签</span></button>
+                            <button type="button" class="btn btn-link btn-lg">电影</button>
+                            <hr>
+                            <button type="button" class="btn btn-link btn-lg">音乐</button>
+                            <hr>
+                            <button type="button" class="btn btn-link btn-lg">python</button>
+                            <hr>
+                            <button type="button" class="btn btn-link btn-lg">感悟</button>
+                            <hr>
+                        </div>
+                    </form>
+                </div>
+            </div>
+            <div class="col-sm-9">
+                <div class="list-group">
+                    <a href="#" class="list-group-item active">
+                      {% for item in items %}
+                        <h5 class="list-group-item-heading"><a href="{{ item[0] }}">{{item[1]}}</a></h5>
+                        <!-- <p class="list-group-item-text">Lorem ipsum dolor sit amet,consectetur adipiscing elit. Aenean</p> -->
+                        <hr>
+                      {% endfor %}
+                    </a>
+                </div>
+            </div>
+        </div>
+</body>
+</html>
