+++ app/templates/regist.html
