--- exts.py
@@ -1,6 +1,0 @@
-# encoding:utf-8
-# 存放db
-
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
