--- templates/login.html
@@ -1,149 +1,0 @@
-
-﻿{% extends "base.html" %}
-
-{% import "bootstrap/wtf.html" as wtf %}
-
-{% block title %}Goodwriter - Sign in{% endblock %}
-
-{% block page_content %}
-
-  <head>
-    <title>Goodwriter</title>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
-    <meta http-equiv="content-type" content="text/html; charset=utf-8"/>
-    <meta name="apple-mobile-web-app-capable" content="yes"/>
-    <link href="resources/css/jquery-ui-themes.css" type="text/css" rel="stylesheet"/>
-    <link href="resources/css/axure_rp_page.css" type="text/css" rel="stylesheet"/>
-    <link href="data/styles.css" type="text/css" rel="stylesheet"/>
-    <link href="files/登录页面/styles.css" type="text/css" rel="stylesheet"/>
-    <script src="resources/scripts/jquery-1.7.1.min.js"></script>
-    <script src="resources/scripts/jquery-ui-1.8.10.custom.min.js"></script>
-    <script src="resources/scripts/axure/axQuery.js"></script>
-    <script src="resources/scripts/axure/globals.js"></script>
-    <script src="resources/scripts/axutils.js"></script>
-    <script src="resources/scripts/axure/annotation.js"></script>
-    <script src="resources/scripts/axure/axQuery.std.js"></script>
-    <script src="resources/scripts/axure/doc.js"></script>
-    <script src="data/document.js"></script>
-    <script src="resources/scripts/messagecenter.js"></script>
-    <script src="resources/scripts/axure/events.js"></script>
-    <script src="resources/scripts/axure/recording.js"></script>
-    <script src="resources/scripts/axure/action.js"></script>
-    <script src="resources/scripts/axure/expr.js"></script>
-    <script src="resources/scripts/axure/geometry.js"></script>
-    <script src="resources/scripts/axure/flyout.js"></script>
-    <script src="resources/scripts/axure/ie.js"></script>
-    <script src="resources/scripts/axure/model.js"></script>
-    <script src="resources/scripts/axure/repeater.js"></script>
-    <script src="resources/scripts/axure/sto.js"></script>
-    <script src="resources/scripts/axure/utils.temp.js"></script>
-    <script src="resources/scripts/axure/variables.js"></script>
-    <script src="resources/scripts/axure/drag.js"></script>
-    <script src="resources/scripts/axure/move.js"></script>
-    <script src="resources/scripts/axure/visibility.js"></script>
-    <script src="resources/scripts/axure/style.js"></script>
-    <script src="resources/scripts/axure/adaptive.js"></script>
-    <script src="resources/scripts/axure/tree.js"></script>
-    <script src="resources/scripts/axure/init.temp.js"></script>
-    <script src="files/登录页面/data.js"></script>
-    <script src="resources/scripts/axure/legacy.js"></script>
-    <script src="resources/scripts/axure/viewer.js"></script>
-    <script src="resources/scripts/axure/math.js"></script>
-    <script type="text/javascript">
-      $axure.utils.getTransparentGifPath = function() { return 'resources/images/transparent.gif'; };
-      $axure.utils.getOtherPath = function() { return 'resources/Other.html'; };
-      $axure.utils.getReloadPath = function() { return 'resources/reload.html'; };
-    </script>
-  </head>
-  <body>
-    <div id="base" class="page-header">
-      {{ wtf.quick_form(form) }}
-      <!-- Unnamed (组合) -->
-      <div id="u145" class="ax_default">
-        <!-- Unnamed (形状) -->
-        <div id="u148" class="ax_default icon">
-          <!-- Unnamed () -->
-          <div id="u149" class="text" style="display: none; visibility: hidden">
-          </div>
-        </div>
-      </div>
-      <!-- Unnamed (矩形) -->
-      <div id="u152" class="ax_default line">
-        <div id="u152_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u153" class="text" style="display: none; visibility: hidden">
-          <p><span></span></p>
-        </div>
-      </div>
-
-      <!-- Unnamed (矩形) -->
-      <div id="u154" class="ax_default box_3">
-        <div id="u154_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u155" class="text" style="visibility: visible;">
-          <p><span>&nbsp;&nbsp; &nbsp;&nbsp; </span><span style="color:#000000;">Please log in Goodwriter</span></p>
-        </div>
-      </div>
-
-      <!-- Unnamed (矩形) -->
-      <div id="u156" class="ax_default box_3">
-        <div id="u156_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u157" class="text" style="visibility: visible;">
-          <p><span>Email</span></p>
-        </div>
-      </div>
-
-      <!-- Unnamed (文本框) -->
-      <div id="u158" class="ax_default text_field">
-        <input id="u158_input" type="text" value=""/>
-      </div>
-
-      <!-- Unnamed (矩形) -->
-      <div id="u159" class="ax_default box_3">
-        <div id="u159_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u160" class="text" style="visibility: visible;">
-          <p>Password</p>
-        </div>
-      </div>
-
-      <!-- Unnamed (文本框) -->
-      <div id="u161" class="ax_default text_field">
-        <input id="u161_input" type="text" value=""/>
-      </div>
-
-      <!-- Unnamed (矩形) -->
-      <div id="u162" class="ax_default primary_button">
-        <div id="u162_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u163" class="text" style="visibility: visible;">
-          <p><span>Log in</span></p>
-        </div>
-      </div>
-
-      <!-- Unnamed (复选框) -->
-      <div id="u164" class="ax_default checkbox">
-        <label for="u164_input">
-          <!-- Unnamed () -->
-          <div id="u165" class="text" style="visibility: visible;">
-            <p><span>Remember me</span></p>
-          </div>
-        </label>
-        <input id="u164_input" type="checkbox" value="checkbox"/>
-      </div>
-
-      <!-- Unnamed (矩形) -->
-      <div id="u166" class="ax_default box_3">
-        <div id="u166_div" class=""></div>
-        <!-- Unnamed () -->
-        <div id="u167" class="text" style="visibility: visible;">
-          <p><span>&nbsp;&nbsp;&nbsp; Forget Password</span></p>
-          <div>
-          <p>New user? <a href="{{ url_for('auth.register') }}">Click here to register</a>.</p>
-          </div>
-        </div>
-      </div>
-    </div>
-
-{% endblock %}
