--- templates/register.html
@@ -1,13 +1,0 @@
-{% extends "base.html" %}
-{% import "bootstrap/wtf.html" as wtf %}
-
-{% block title %}Goodwriter - Sign up{% endblock %}
-
-{% block page_content %}
-<div class="page-header">
-    <h1>Sign up</h1>
-</div>
-<div class="col-md-4">
-    {{ wtf.quick_form(form) }}
-</div>
-{% endblock %}
